# Make IntMap::new const