Changes:

- `IntMap::new` is now const, thus it doesn't allocate memory.
- `IntMap::with_capacity` doesn't allocate memory if capacity is zero. Previously it allocated space for at least one element.
- Previously we assumed that there is at least space for one more element. Therefore `IntMap::insert` and `IntMap::insert_checked` didn't need to increase the cache at the beginning. This changed. I added another field `IntMap::needs_increase` to keep track if an increase is necessary. Now we handle this flag at the beginning of these functions and set it at the end.
- All getter functions need to handle an empty `IntMap` properly. I added shortcuts at the beginning of these functions.

Closed #49