@JesperAxelsson 2024-08-02
Regarding the inserts, the plan was to reduce the number of branches as much as possible. Being able to use it in `const` context is more useful and couldn't detect any performance reduction from the benchmarks. 

Would be great if you could rebase on master for the CI fixes :)