@JesperAxelsson 2024-08-03
This seems to be a failed test this time. For some reason the ordering of debug print has changed.