@jakoschiko 2024-08-03
Yes, I'll investigate this later. I'm not yet done with the implementation, thus this PR is on draft. First I'll need to think about some edge cases.