@jakoschiko 2024-08-03
I'd would like to add more integration tests using `proptest`. Should I add them to `./tests` or create another folder and crate in `./integration_tests`?