@JesperAxelsson 2024-08-03
Create a new crate is the cleaner option I would say.

At the moment it look like we perform two checks for load rate. One at the start of insert and one after. Might be better to only ensure load rate the beginning with count + 1.  The downside of this approach is that sometimes we reallocate even if we don't need. In theory at least. Size increase with the power of 2 so we rarely actually need to increase the size. 
Still might cause issues if you set a capacity and load rate but we still reallocate when we are on the capacity. Unless of course the initial capacity takes the extra element into account.

Just some thoughts, feels like I'm mind dumping a bit ^^