@jakoschiko 2024-08-03
> Create a new crate is the cleaner option I would say.

IMO this is a little bit unusual. But no problem, will do it. I'll create a separate PR for that.

> At the moment it look like we perform two checks for load rate. One at the start of insert and one after. Might be better to only ensure load rate the beginning with count + 1. The downside of this approach is that sometimes we reallocate even if we don't need. In theory at least. Size increase with the power of 2 so we rarely actually need to increase the size. Still might cause issues if you set a capacity and load rate but we still reallocate when we are on the capacity. Unless of course the initial capacity takes the extra element into account.

I had similar thoughts. Though the second check is not really a check, more like a postponement of the actual check. And it can be implemented branchless because it's only about setting the `bool`. However, I'm not feeling well making changes to these functions before adding more `proptest` tests.

> Just some thoughts, feels like I'm mind dumping a bit ^^

Always happy about that. Thanks!