@jakoschiko 2024-08-11
I'm happy with the current state.

Benchmark results:

master (3668ca25d94a)

```
test tests::u64_get_built_in                     ... bench:      80,048.49 ns/iter (+/- 1,120.25)
test tests::u64_get_indexmap                     ... bench:     104,913.54 ns/iter (+/- 707.30)
test tests::u64_get_intmap                       ... bench:      12,750.83 ns/iter (+/- 102.93)
test tests::u64_insert_built_in                  ... bench:     108,822.23 ns/iter (+/- 434.90)
test tests::u64_insert_built_in_without_capacity ... bench:     282,702.10 ns/iter (+/- 2,019.04)
test tests::u64_insert_indexmap                  ... bench:     136,701.40 ns/iter (+/- 882.86)
test tests::u64_insert_intmap                    ... bench:      44,715.89 ns/iter (+/- 707.91)
test tests::u64_insert_intmap_checked            ... bench:      42,648.63 ns/iter (+/- 1,921.58)
test tests::u64_insert_intmap_entry              ... bench:      41,845.32 ns/iter (+/- 967.49)
test tests::u64_insert_intmap_without_capacity   ... bench:     509,396.55 ns/iter (+/- 3,110.16)
test tests::u64_resize_intmap                    ... bench:      31,002.43 ns/iter (+/- 79.86)
```

PR branch (82eaac8da302)

```
running 11 tests
test tests::u64_get_built_in                     ... bench:      75,040.76 ns/iter (+/- 614.64)
test tests::u64_get_indexmap                     ... bench:      99,090.88 ns/iter (+/- 703.46)
test tests::u64_get_intmap                       ... bench:      11,704.59 ns/iter (+/- 149.24)
test tests::u64_insert_built_in                  ... bench:     108,497.07 ns/iter (+/- 206.74)
test tests::u64_insert_built_in_without_capacity ... bench:     262,929.10 ns/iter (+/- 679.97)
test tests::u64_insert_indexmap                  ... bench:     118,672.51 ns/iter (+/- 5,870.27)
test tests::u64_insert_intmap                    ... bench:      34,308.19 ns/iter (+/- 413.66)
test tests::u64_insert_intmap_checked            ... bench:      37,221.59 ns/iter (+/- 551.35)
test tests::u64_insert_intmap_entry              ... bench:      59,495.75 ns/iter (+/- 522.95)
test tests::u64_insert_intmap_without_capacity   ... bench:     486,095.35 ns/iter (+/- 2,913.17)
test tests::u64_resize_intmap                    ... bench:      29,261.58 ns/iter (+/- 152.12)
```